requirements.txt
onnxruntime==1.16.3
numpy==1.24.3
Pillow==10.0.1